+++ python/ray/serve/tests/test_api.py
@@ -10,11 +10,8 @@ import starlette.responses
 import ray
 from ray import serve
 from ray.test_utils import wait_for_condition
-from ray.serve.constants import SERVE_PROXY_NAME
-from ray.serve.exceptions import RayServeException
 from ray.serve.config import BackendConfig
-from ray.serve.utils import (format_actor_name, get_all_node_ids,
-                             get_random_letters)
+from ray.serve.utils import get_random_letters
 
 
 def test_e2e(serve_instance):
@@ -501,48 +498,6 @@ def test_endpoint_input_validation(serve_instance):
     with pytest.raises(TypeError):
         serve.create_endpoint("endpoint", backend=2)
     serve.create_endpoint("endpoint", backend="backend")
-
-
-def test_shutdown():
-    def f():
-        pass
-
-    serve.start(http_port=8003)
-    serve.create_backend("backend", f)
-    serve.create_endpoint("endpoint", backend="backend")
-
-    actor_names = [
-        serve.api._global_client._controller_name,
-        format_actor_name(SERVE_PROXY_NAME,
-                          serve.api._global_client._controller_name,
-                          get_all_node_ids()[0][0])
-    ]
-
-    def check_alive():
-        alive = True
-        for actor_name in actor_names:
-            try:
-                ray.get_actor(actor_name)
-            except ValueError:
-                alive = False
-        return alive
-
-    wait_for_condition(check_alive)
-
-    serve.shutdown()
-    with pytest.raises(RayServeException):
-        serve.list_backends()
-
-    def check_dead():
-        for actor_name in actor_names:
-            try:
-                ray.get_actor(actor_name)
-                return False
-            except ValueError:
-                pass
-        return True
-
-    wait_for_condition(check_dead)
 
 
 def test_shadow_traffic(serve_instance):
